refactor tests to internal/policy
